+++ tests/test_version_info.py
@@ -1,0 +1,11 @@
+import os
+
+from laktory import show_version_info
+
+
+def test_show_version_info():
+    print(show_version_info())
+
+
+if __name__ == "__main__":
+    test_show_version_info()